Adding test postgres container for unit tests

+++ customer/src/main/java/com/aplazo/customer/controllers/CustomerController.java
@@ -5,7 +5,6 @@ import com.aplazo.customer.services.CustomerService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import java.math.BigInteger;
 import java.util.NoSuchElementException;
 
 @RestController
@@ -15,7 +14,7 @@ public class CustomerController {
     private final CustomerService customerService;
 
     @GetMapping("/{id}")
-    public Customer findById(@PathVariable BigInteger id) {
+    public Customer findById(@PathVariable Long id) {
         return customerService.findById(id).orElseThrow( ()->
             {
                 throw new NoSuchElementException("Customer with ID " + id + " does not exist.");
